backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
httpx
openmeteo-requests
requests-cache
retry-requests